fix tests. Introduce builder pattern as utility class for creation of Evalution for tests.
I'm focusing on Evalution since this part is easy to understand for me.

So I am skipping the part of client, standards etc. To focus on it letter - now it is just to many complecity - and I want to focus on AAssesment as soon as I can.

But lets make some evalution logic:

9. The assessment has an expiration date of 365 days counting from the day
evaluation took place. After it is exceeded, the assessment expires.